+++ linux/services/mqtt-subscriptions/main.cpp
@@ -14,11 +14,11 @@ void readMultipleRadioNodes(mqtt::async_client& mqtt_client)
     CommandCallback commandCallback(mqtt_client);
 
     mqtt_client.set_callback(commandCallback);
-    mqtt_client.subscribe("radio-arduino/DDATA/#", QOS)->wait();
-    mqtt_client.subscribe("radio-arduino/DBIRTH/#", QOS)->wait();
-    mqtt_client.subscribe("radio-arduino/ADVERTISEMENT/#", QOS)->wait();
-    mqtt_client.subscribe("radio-arduino/STATE/#", QOS)->wait();
-    mqtt_client.subscribe("raduino-subscriptions/RCMD/#", QOS)->wait();
+    mqtt_client.subscribe("raduino-bridge/DDATA/#", QOS)->wait();
+    mqtt_client.subscribe("raduino-bridge/DBIRTH/#", QOS)->wait();
+    mqtt_client.subscribe("raduino-bridge/ADVERTISEMENT/#", QOS)->wait();
+    mqtt_client.subscribe("raduino-bridge/STATE/#", QOS)->wait();
+    mqtt_client.subscribe("raduino-subscription/RCMD/#", QOS)->wait();
 
     commandCallback.resendBirthCertificate();